Fix timestamp handling in MDTM
There were two problems with the timestamp handling with MDTM:

1. In vsf_sysutil_parse_time(), the `the_time.tm_isdst` attribute was
   always set to 0, regardless of whether DST (daylight saving time)
   is active on the given date or not.

   This made glibc shift the timestamp when DST was in fact active on
   the given date, in an attempt to correct the discrepancy between
   the given timestamp and the `tm_isdst` attribute. The shifting
   produced incorrect results however.

   We fix this by setting `tm_isdst` to -1 to let glibc decide if DST
   is active or not at the time of the timestamp. glibc won't touch
   the timestamp then.

2. vsftpd used to record the offset from UTC of the current timezone
   in the global variable `s_timezone`. This variable was then
   subtracted from the variable `the_time` in vsf_sysutil_setmodtime()
   when the config option use_localtime=NO was set. This was done to
   compensate for the fact that mktime(), used in
   vsf_sysutil_parse_time(), expects a timestamp expressed as local
   time, whereas vsftpd is dealing with universal time.

   However, this did not work in the case when the offset stored in
   `s_timezone` did not match the timezone of the timestamp given to
   mktime() - this happens when DST is active at the current time, but
   DST is not active at the time of the timestamp, or vice versa.

   We fix this by subtracting the real timezone offset directly in
   vsf_sysutil_parse_time().

   Note that the `tm_gmtoff` attribute, used in this fix, is a
   BSD/glic extension. However, using `tm_gmtoff` seems like the
   simplest solution and we need to make this work only with glibc
   anyway.

The fix was tested in the following way. We checked that the timestamp
given to the MDTM command when setting modification time exactly
matches the timestamp received as response from MDTM when reading back
the modification time. Additionally, we checked that the modification
time was set correctly on the given file on disk.

These two checks were performed under various conditions - all the
combinations of DST/non-DST system time, DST/non-DST modification
time, use_localtime=YES/NO.

Note that (I think) this will still not work if the rules for when DST
is active change. For example, if DST is ever completely cancelled in
the Europe/Prague timezone, and vsftpd is dealing with a timestamp
from a time when DST was active, it will produce incorrect results. I
think we would need the full zone file to fix this, but the zone file
is hard to provide when we're chroot-ed.

Resolves: rhbz#1567855